docs: trim PR narration and issue-number refs from comments (#1078)
Addresses Copilot review feedback on PR #1113: drop file-shape-cap
motivation from module comments (rule 05 wants purpose/usage, not
refactor narration), remove an issue-number reference from a module
doc, collapse a duplicated multi-line rustdoc block to one line, and
move suggestion_dropdown.rs's four non-trivial tests to a sibling
tests.rs per the sibling-file convention.
