+++ frontend/src/components/mod.rs
@@ -45,12 +45,11 @@ pub mod chip_editor;
 pub use chip_editor::ChipEditor;
 
 // Autocomplete-dropdown machinery (selection state, the dropdown component,
-// the substring filter) shared by `chip_editor` and `suggest_field` — split
-// out of `chip_editor.rs` once the file crossed the file-shape cap.
+// the substring filter) shared by `chip_editor` and `suggest_field`.
 mod suggestion_dropdown;
 
-// Single-value sibling of `chip_editor`'s dropdown (e.g. the Series field) —
-// split out of `chip_editor.rs` once the file crossed the file-shape cap.
+// Single-value sibling of `chip_editor`'s dropdown, for fields that hold one
+// value rather than a list (e.g. the Series field).
 pub mod suggest_field;
 pub use suggest_field::SuggestField;
