+++ frontend/src/components/suggestion_dropdown/mod.rs
@@ -8,6 +8,9 @@ use dioxus::prelude::*;
 
 use super::chip_editor::SuggestionItem;
 
+#[cfg(test)]
+mod tests;
+
 /// Cap on suggestion rows. Five fits without scroll on a typical
 /// edit-row layout and matches the "< 5 relevant suggestions"
 /// user requirement.
@@ -145,41 +148,5 @@ pub(super) fn compute_suggestions(
             .take(MAX_SUGGESTIONS)
             .cloned()
             .collect()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn compute_suggestions_returns_empty_when_pool_is_empty() {
-        let result = compute_suggestions(&[], &[], "", true, false);
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn compute_suggestions_filters_already_chosen_values() {
-        let pool = vec![
-            SuggestionItem::new("Ada", 1),
-            SuggestionItem::new("Mira", 2),
-        ];
-        let result = compute_suggestions(&pool, &["Ada".to_string()], "", true, false);
-        assert_eq!(result.len(), 1);
-        assert_eq!(result[0].name, "Mira");
-    }
-
-    #[test]
-    fn compute_suggestions_returns_empty_when_unfocused_and_query_empty() {
-        let pool = vec![SuggestionItem::new("Ada", 1)];
-        let result = compute_suggestions(&pool, &[], "", false, false);
-        assert!(result.is_empty());
-    }
-
-    #[test]
-    fn compute_suggestions_returns_empty_when_suppress_open_is_true() {
-        let pool = vec![SuggestionItem::new("Ada", 1)];
-        let result = compute_suggestions(&pool, &[], "", true, true);
-        assert!(result.is_empty());
     }
 }
